requirements.txt
anyio==4.12.1
appnope==0.1.4
asttokens==3.0.1
boto3==1.42.23
botocore==1.42.23
certifi==2026.1.4
charset-normalizer==3.4.4
comm==0.2.3
debugpy==1.8.19
decorator==5.2.1
dotenv==0.9.9
executing==2.2.1
faiss-cpu==1.13.2
filelock==3.20.2
fsspec==2025.12.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
jupyter_client==8.7.0
jupyter_core==5.9.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.0
packaging==25.0
parso==0.8.5
pexpect==4.9.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pypdf==6.5.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyzmq==27.1.0
requests==2.32.5
s3transfer==0.16.0
semanticscholar==0.11.0
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
tenacity==9.1.2
torch==2.9.1
tornado==6.5.4
traitlets==5.14.3
typing_extensions==4.15.0
urllib3==2.6.2
wcwidth==0.2.14
